backend-app/requirements.txt
Django==1.11.6
djangorestframework==3.7.1
psycopg2==2.7.3.2
PyJWT==1.5.3
djangorestframework-jwt==1.11.0
